Removed some compare unsigned warnings

+++ HOPSAN++/HopsanGUI/Dialogs/ComponentPropertiesDialog.cpp
@@ -98,12 +98,12 @@ void ComponentPropertiesDialog::createEditStuff()
         //Adjust sizes of labels, to make sure that all text is visible and that the spacing is not too big between them
     int descriptionSize=30;
     int nameSize = 10;
-    for(size_t i=0; i<mvParameterLayout.size(); ++i)
+    for(int i=0; i<mvParameterLayout.size(); ++i)
     {
         descriptionSize = std::max(descriptionSize, mvParameterLayout.at(i)->mDescriptionNameLabel.width());
         nameSize = std::max(nameSize, mvParameterLayout.at(i)->mDataNameLabel.width());
     }
-    for(size_t i=0; i<mvParameterLayout.size(); ++i)
+    for(int i=0; i<mvParameterLayout.size(); ++i)
     {
         mvParameterLayout.at(i)->mDescriptionNameLabel.setFixedWidth(descriptionSize+10);   //Offset of 10 as extra margin
         mvParameterLayout.at(i)->mDataNameLabel.setFixedWidth(nameSize+10);
@@ -157,17 +157,17 @@ void ComponentPropertiesDialog::createEditStuff()
 
     descriptionSize=30;
     nameSize = 10;
-    for(size_t j=0; j<mvStartValueLayout.size(); ++j)
+    for(int j=0; j<mvStartValueLayout.size(); ++j)
     {
-        for(size_t i=0; i<mvStartValueLayout.at(j).size(); ++i)
+        for(int i=0; i<mvStartValueLayout.at(j).size(); ++i)
         {
             descriptionSize = std::max(descriptionSize, mvStartValueLayout.at(j).at(i)->mDescriptionNameLabel.width());
             nameSize = std::max(nameSize, mvStartValueLayout.at(j).at(i)->mDataNameLabel.width());
         }
     }
-    for(size_t j=0; j<mvStartValueLayout.size(); ++j)
+    for(int j=0; j<mvStartValueLayout.size(); ++j)
     {
-        for(size_t i=0; i<mvStartValueLayout.at(j).size(); ++i)
+        for(int i=0; i<mvStartValueLayout.at(j).size(); ++i)
         {
             mvStartValueLayout.at(j).at(i)->mDescriptionNameLabel.setFixedWidth(descriptionSize+10);   //Offset of 10 as extra margin
             mvStartValueLayout.at(j).at(i)->mDataNameLabel.setFixedWidth(nameSize+10);
